requirements.txt
torch
torchvision
pytesseract
opencv-python
pdf2image
python-Levenshtein